added tests for tag controller/service and board service

+++ server/src/main/java/server/Services/BoardService.java
@@ -40,6 +40,8 @@ public class BoardService {
 
     public IdResponseModel deleteBoard(int boardId) {
         try {
+            if (!boardRepository.existsById(boardId))
+                return new IdResponseModel(-1, "Board does not exist");
             boardRepository.deleteById(boardId);
             return new IdResponseModel(boardId, null);
         } catch (Exception e) {
